+++ tests/test_oauth2_authorization_code.py
@@ -3,7 +3,7 @@ import pytest
 import requests
 
 import requests_auth
-from tests.oauth2_helper import token_cache, browser_mock, BrowserMock
+from requests_auth.testing import BrowserMock, browser_mock, token_cache
 from tests.auth_helper import get_header, get_request